backend/notifier/requirements.txt
apprise
loguru
paho-mqtt==1.6.1
pydantic
pydantic-settings